Make ticktock worker callbacks configurable
Previously after processing message or when error was raised during
processing of a message, there were explicit calls to Honeybadger and
MonitoredScheduledTask.

This change allows the same functionality to be used by configuring the
SneakersToolbox, freeing the user being bound to these constants being
defined or forcing checks for their presence in toolbox code.